# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API para manejar stocks y recomendaciones de inversión.",
        "title": "Stock Ratings API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/",
    "paths": {
        "/stocks": {
            "get": {
                "description": "Retorna la lista completa de acciones disponibles en la base de datos.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stocks"
                ],
                "summary": "Obtiene todos los stocks",
                "responses": {
                    "200": {
                        "description": "Lista de acciones disponibles",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Stock"
                            }
                        }
                    }
                }
            }
        },
        "/stocks/recommendation": {
            "get": {
                "description": "Retorna la acción con mayor potencial de crecimiento según los datos almacenados. Si no hay recomendaciones, devuelve `null`.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stocks"
                ],
                "summary": "Obtiene la mejor inversión",
                "responses": {
                    "200": {
                        "description": "La mejor acción recomendada",
                        "schema": {
                            "$ref": "#/definitions/models.Stock"
                        }
                    }
                }
            }
        },
        "/stocks/sync": {
            "post": {
                "description": "Llama a un servicio externo para actualizar los datos de los stocks. Requiere un token de autorización en el header.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Stocks"
                ],
                "summary": "Sincroniza los datos de los stocks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Token de autorización (Bearer)",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    },
    "definitions": {
        "models.Stock": {
            "description": "Modelo de una acción en la base de datos",
            "type": "object",
            "properties": {
                "action": {
                    "type": "string",
                    "example": "Buy"
                },
                "brokerage": {
                    "type": "string",
                    "example": "Goldman Sachs"
                },
                "company": {
                    "type": "string",
                    "example": "Apple Inc."
                },
                "created_at": {
                    "type": "string",
                    "example": "2025-03-02T15:04:05Z"
                },
                "deleted_at": {
                    "type": "string",
                    "example": "2025-03-02T15:04:05Z"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "rating_from": {
                    "type": "string",
                    "example": "Neutral"
                },
                "rating_to": {
                    "type": "string",
                    "example": "Buy"
                },
                "target_from": {
                    "type": "number",
                    "example": 150
                },
                "target_to": {
                    "type": "number",
                    "example": 180
                },
                "ticker": {
                    "type": "string",
                    "example": "AAPL"
                },
                "time": {
                    "type": "string",
                    "example": "2025-03-02"
                },
                "updated_at": {
                    "type": "string",
                    "example": "2025-03-02T15:04:05Z"
                }
            }
        }
    }
}